test: interactions for the update prompt
